fix: 削除機能の修正

+++ src/models/knowledge.repository.ts
@@ -2,6 +2,7 @@ import { glob, readFile, unlink, writeFile } from 'node:fs/promises';
 
 import type { Knowledge } from './knowledge.model.js';
 
+// knowledgeの一覧を取得する
 async function getAll(): Promise<Knowledge[]> {
   const files = await Array.fromAsync(glob('./storage/**/*.json'));
 
@@ -9,6 +10,7 @@ async function getAll(): Promise<Knowledge[]> {
 
   return knowledges;
 }
+// 新規作成の保存
 async function upsert(knowledge: Knowledge): Promise<void> {
   const create_file = `./storage/${knowledge.knowledgeId}.json`;
   const data = JSON.stringify(knowledge);
@@ -20,15 +22,24 @@ const uuidPattern = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}
 function isKnowledgeId(value: string): boolean {
   return uuidPattern.test(value);
 }
-
-async function deleteByKnowledgeId(knowledgeId: string): Promise<void> {
+// knowledgeを削除
+async function deleteByKnowledgeId(knowledgeId: string, userId: string): Promise<boolean> {
   console.log(`Deleting knowledge with knowledgeId: ${knowledgeId}`);
   if (!isKnowledgeId(knowledgeId)) {
     throw new Error('Invalid knowledgeId');
   }
-
+  let knowledge: Knowledge;
   const filePath = `./storage/${knowledgeId}.json`;
+  try {
+    knowledge = JSON.parse(await readFile(filePath, 'utf-8'));
+  } catch {
+    return false;
+  }
+  if (userId !== knowledge.authorId) {
+    return false;
+  }
   await unlink(filePath);
+  return true;
 }
 
 export const KnowledgeRepository = {
